training: move metrics and progress handling from Job to Task
This is more robust in the face of crashes from the master daemon
